fix: 植物种植的时候能够减少阳光了

+++ Assets/Scripts/ScriptableObjeect/PlantSO.cs
@@ -14,5 +14,5 @@ public class PlantSO : ScriptableObject
 
     [Header("Basic Params")]
     public float cooldownTimerMax;
-    public float sunCost;
+    public int sunCost;
 }
